Factor out SB upgrades stuff
